Fix unused variables in Analysis refactor
