+++ web/shapeworks/src/components/Analysis/util.ts
@@ -1,0 +1,5 @@
+export enum AnalysisTabs {
+    PCA = 0,
+    Groups,
+    Charts
+}
